Update azure-pipelines.yml for Azure Pipelines (#48)
This pull request makes a small change to the `azure-pipelines.yml`
build pipeline to improve visibility and reliability when setting up
`corepack`. The change adds logging of the `corepack` version before and
after installation, and ensures the latest version is installed.
PRimarily the corepack update is there to handle key rotation issues.

* Build pipeline setup:
* Logs the `corepack` version before and after installation, and
installs the latest version globally to ensure the build uses the most
up-to-date tooling.